Fixed database
Previously, did not recognize if a table had already been created or not between sessions.
Now DBM is a singleton.  It manages all of the tables and there is only one of them.
When it is constructed (when program starts) it checks the database.
It marks all tables that already exist, puts them into a map, and returns those tables instead.
Unofficaly tested using the login screen. UN and PW was stored, and able to be verified between sessions.
Tests were also re-written and they all work as expected.